requirements.txt
pySerial==3.4
